Move weather-based rain delay into it's own adjustment method
The previous implementation caused the scale to change to zero for
user’s not using a weather adjustment method. This is not ideal since
the water scale will remain at 0% until there is user intervention.

Now it will be a user selectable method which uses a user-configurable
delay value which defaults to 24 hours.